be/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
tortoise-orm
passlib[argon2]
authlib
pywebpush
python-multipart